pyproject.toml
[tool.poetry]
name = "django-ghostpost"
version = "0.1.0"
description = ""
authors = ["Mike A. <jimalik2525@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.6"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.1"
autopep8 = "^1.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
